requirements.txt
nltk==3.9.1
pandas==2.2.2
Pillow==10.4.0
streamlit==1.37.1
torch==2.4.0
transformers==4.43.3
